Add files via upload
The Online Exam Management System is a web-based application that facilitates the secure and efficient conduction of exams in an academic setting. This system is designed to offer a comprehensive solution for educators and institutions to create, manage, and administer exams, while providing students with a user-friendly interface for taking exams online.